server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
prisma==0.13.1
pydantic-settings==2.3.4
google-cloud-storage==2.17.0
httpx==0.27.0
cryptography==42.0.8
pypdf==4.2.0
langchain-core==1.2.18
langchain-openai==1.1.11
Pillow==10.4.0
